Fixed issues pointed out by copilot
